feat(utils/throttler): throttler revamped (#3750)
* refactor: move throttler type into its own subPackage

* refactor(utils/throttler): swap setters for the option pattern

* refactor(utils/throttler): max queue len to uint64

* chore(throttler): add context to calls

* fix(node): test compilation error

* test(throttler): use synctest

* fix(throtler): queue 0 behaves 0 and no longer forbids concurrent
executions

* chore(throttler): make everything a uint64

* fix(throttler): typo

+++ node/node.go
@@ -301,7 +301,7 @@ func New(cfg *Config, version string, logLevel *log.Level) (*Node, error) {
 			logger,
 		),
 		cfg.MaxConcurrentCompilations,
-		int32(cfg.MaxCompilationQueue),
+		uint64(cfg.MaxCompilationQueue),
 	)
 
 	if cfg.Sequencer {
@@ -320,7 +320,7 @@ func New(cfg *Config, version string, logLevel *log.Level) (*Node, error) {
 			FeeTokenAddresses: feeTokens,
 		}
 		nodeVM = vm.New(&chainInfo, false, logger)
-		throttledVM = NewThrottledVM(nodeVM, cfg.MaxVMs, int32(cfg.MaxVMQueue))
+		throttledVM = NewThrottledVM(nodeVM, cfg.MaxVMs, uint64(cfg.MaxVMQueue))
 		mempool := mempool.New(database, chain, mempoolLimit, logger)
 		executor := builder.NewExecutor(chain, nodeVM, logger, cfg.SeqDisableFees, false)
 		builder := builder.New(chain, executor)
@@ -380,7 +380,7 @@ func New(cfg *Config, version string, logLevel *log.Level) (*Node, error) {
 			FeeTokenAddresses: feeTokens,
 		}
 		nodeVM = vm.New(&chainInfo, false, logger)
-		throttledVM = NewThrottledVM(nodeVM, cfg.MaxVMs, int32(cfg.MaxVMQueue))
+		throttledVM = NewThrottledVM(nodeVM, cfg.MaxVMs, uint64(cfg.MaxVMQueue))
 
 		feederGatewayDataSource := sync.NewFeederGatewayDataSource(chain, adaptfeeder.New(client))
 		synchronizer = sync.New(
